+++ packages/plugins/operators/operators-js/src/operators/shared/intl.js
@@ -55,6 +55,6 @@ function intl({ params, location, methodName }) {
   });
 }
 
-intl.dynamic = false;
+intl.dynamic = true;
 
 export default intl;