+++ packages/utils/block-dev-e2e/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lowdefy/block-dev-e2e
 
+## 4.7.3
+
+### Patch Changes
+
+- @lowdefy/e2e-utils@4.7.3
+
 ## 4.7.2
 
 ### Patch Changes