+++ packages/utils/block-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.7.3
+
+### Patch Changes
+
+- @lowdefy/errors@4.7.3
+- @lowdefy/helpers@4.7.3
+
 ## 4.7.2
 
 ### Patch Changes