+++ packages/utils/e2e-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lowdefy/e2e-utils
 
+## 4.7.3
+
+### Patch Changes
+
+- @lowdefy/helpers@4.7.3
+
 ## 4.7.2
 
 ### Patch Changes